Nucleus: Remove `Academics` from Channeli apps and move the new link to `Links` dir
XXX: Decision to move this was taken as IMG won't be doing Academics work from
now on and as the acad data as well the apps are no longer maintained by us, its
better to remove the links.

For easier usage and accountability for Academics site, it has been added in the
links(as an external link).

\cc @Namit @Harshit @Jeet